+++ src/components/AlphabetFilter/alphabetFilterChangelogData.ts
@@ -10,6 +10,13 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "2025-02-13",
+    version: "3.5.4",
+    type: "Update",
+    affects: ["Styles"],
+    notes: ["Removed bold font weight."],
+  },
+  {
     date: "2024-03-14",
     version: "3.0.0",
     type: "Update",